requirements.txt
pytesseract
Pillow
scikit-learn
pandas
joblib
spacy
pymupdf
